Only load the models once
